+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-16 14:50:50
-// Commit: perf(core): apply RepaintBoundary on animated core widgets
+// Updated at: 2026-04-16 15:53:47
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '422';
-  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
-  final DateTime timestamp = DateTime.parse('2026-04-16T14:50:50+05:00');
+  final String commitRef = '427';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-04-16T15:53:47+05:00');
 }